Implement collection assets as a core property of collections
Fixes #266
