+++ tests/data-files/collections/with_assets.json
@@ -1,0 +1,167 @@
+{
+    "stac_version": "1.0.0-beta.2",
+    "stac_extensions": [],
+    "id": "landsat-8-l1",
+    "title": "Landsat 8 L1",
+    "description": "Landat 8 imagery radiometrically calibrated and orthorectified using gound points and Digital Elevation Model (DEM) data to correct relief displacement.",
+    "keywords": [
+        "landsat"
+    ],
+    "extent": {
+        "spatial": {
+            "bbox": [
+                [
+                    -180,
+                    -90,
+                    180,
+                    90
+                ]
+            ]
+        },
+        "temporal": {
+            "interval": [
+                [
+                    "2013-06-01T00:00:00Z",
+                    null
+                ]
+            ]
+        }
+    },
+    "providers": [
+        {
+            "name": "USGS",
+            "url": "https://landsat.usgs.gov/",
+            "roles": [
+                "producer",
+                "licensor"
+            ]
+        },
+        {
+            "name": "Planet Labs",
+            "url": "https://github.com/landsat-pds/landsat_ingestor",
+            "roles": [
+                "processor"
+            ]
+        },
+        {
+            "name": "AWS",
+            "url": "https://landsatonaws.com/",
+            "roles": [
+                "host"
+            ]
+        },
+        {
+            "name": "Development Seed",
+            "url": "https://developmentseed.org/",
+            "roles": [
+                "processor"
+            ]
+        }
+    ],
+    "license": "PDDL-1.0",
+    "summaries": {
+        "platform": [
+            "landsat-8"
+        ],
+        "instruments": [
+            "oli",
+            "tirs"
+        ],
+        "gsd": [
+            30
+        ],
+        "eo:bands": [
+            {
+                "name": "B1",
+                "common_name": "coastal",
+                "center_wavelength": 0.44,
+                "full_width_half_max": 0.02
+            },
+            {
+                "name": "B2",
+                "common_name": "blue",
+                "center_wavelength": 0.48,
+                "full_width_half_max": 0.06
+            },
+            {
+                "name": "B3",
+                "common_name": "green",
+                "center_wavelength": 0.56,
+                "full_width_half_max": 0.06
+            },
+            {
+                "name": "B4",
+                "common_name": "red",
+                "center_wavelength": 0.65,
+                "full_width_half_max": 0.04
+            },
+            {
+                "name": "B5",
+                "common_name": "nir",
+                "center_wavelength": 0.86,
+                "full_width_half_max": 0.03
+            },
+            {
+                "name": "B6",
+                "common_name": "swir16",
+                "center_wavelength": 1.6,
+                "full_width_half_max": 0.08
+            },
+            {
+                "name": "B7",
+                "common_name": "swir22",
+                "center_wavelength": 2.2,
+                "full_width_half_max": 0.2
+            },
+            {
+                "name": "B8",
+                "common_name": "pan",
+                "center_wavelength": 0.59,
+                "full_width_half_max": 0.18
+            },
+            {
+                "name": "B9",
+                "common_name": "cirrus",
+                "center_wavelength": 1.37,
+                "full_width_half_max": 0.02
+            },
+            {
+                "name": "B10",
+                "common_name": "lwir11",
+                "center_wavelength": 10.9,
+                "full_width_half_max": 0.8
+            },
+            {
+                "name": "B11",
+                "common_name": "lwir12",
+                "center_wavelength": 12,
+                "full_width_half_max": 1
+            }
+        ]
+    },
+    "links": [
+        {
+            "rel": "self",
+            "href": "https://landsat-stac.s3.amazonaws.com/landsat-8-l1/catalog.json"
+        },
+        {
+            "rel": "parent",
+            "href": "https://landsat-stac.s3.amazonaws.com/catalog.json"
+        },
+        {
+            "rel": "root",
+            "href": "https://landsat-stac.s3.amazonaws.com/catalog.json"
+        },
+        {
+            "rel": "child",
+            "href": "https://landsat-stac.s3.amazonaws.com/landsat-8-l1/paths/catalog.json"
+        }
+    ],
+    "assets": {
+        "test": {
+          "href": "./test.json",
+          "title": "A sample asset for testing",
+          "type": "text/json"
+        }
+    }
+}
